add update unit and changed delete with configration

+++ app/Http/Controllers/UnitController.php
@@ -40,16 +40,54 @@ class UnitController extends Controller
     public function destroy(Request $request)
     {
 
+
+        if(is_null( $request->input('unit_id')) || empty( $request->input('unit_id'))){
+
+            Session::flash('message','Unit ID is required');
+            return redirect()->back();
+
+        }
+
         $id=$request->input('unit_id');
         Unit::destroy($id);
 
         Session::flash('message','Unit has been deleted');
-
         return redirect()->back();
 
     }
 
     public function update(Request $request)
+    {
+
+        $request->validate([
+            'unit_code'=>'required',
+            'unit_id'=>'required',
+            'unit_name'=>'required',
+        ]);
+
+        $unitID=intval($request->input('unit_id'));
+
+        $unit=Unit::find($unitID);
+
+        $unit->unit_name=$request->input('unit_name');
+        $unit->unit_code=$request->input('unit_code');
+
+        $unit->save();
+
+
+        Session::flash('message','Unit '. $unit->unit_name .' has been updated');
+        return redirect()->back();
+
+
+
+
+
+
+
+
+    }
+
+    public function search(Request $request)
     {
 
     }
